Safari-611.1.10/LayoutTests/fast/css/style-attribute-invalidation-propagates-to-direct-siblings-expected.txt

Test that style invalidation through inline-style propagates correctly to siblings when a combinator requires it.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS getComputedStyle(document.querySelectorAll("target")[0]).color is "rgb(0, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("target")[1]).color is "rgb(0, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("target")[0]).color is "rgb(255, 192, 203)"
PASS getComputedStyle(document.querySelectorAll("target")[1]).color is "rgb(255, 192, 203)"
PASS getComputedStyle(document.querySelectorAll("target")[0]).color is "rgb(0, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("target")[1]).color is "rgb(0, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("target")[0]).color is "rgb(255, 192, 203)"
PASS getComputedStyle(document.querySelectorAll("target")[1]).color is "rgb(255, 192, 203)"
PASS getComputedStyle(document.querySelectorAll("target")[0]).color is "rgb(255, 192, 203)"
PASS getComputedStyle(document.querySelectorAll("target")[1]).color is "rgb(255, 192, 203)"
PASS successfullyParsed is true

TEST COMPLETE
